Enable the use of the second timer
In previous versions, driving of heatbed used the algorithm from Prusa fw 3.7.2 and with a disabled second timer.
To use all the latest algorithms I had to change 2 timers (`Timer0` -> `Timer4` and `Timer2` -> `Timer0`)
Explanation:
Rambo board uses `Timer0` (`OC0B` register) on pin `PB7` to drive the bed. 
For RAMPS board it has to be switched to `Timer4` (`OC4C` register) as it uses a different pin (`PG5`).
To enable the second timer, Rambo board utilizes `Timer2` (`OC2B` register) on pin `PH6`, which is unused.
This pin, however, is driving the hotend fan on RAMPS boards, so `Timer2` can't be used in my case.
Instead, I replaced `Timer2` with `Timer0` (`OC0B`) on an unused pin `PG5` (or `D4`, which is usually used to drive a servo motor).